+++ src/commands.ts
@@ -8,6 +8,7 @@ import { Fields } from "./form"
 import fortuneURL from "../resources/fortune.txt"
 import { Gate } from './gate'
 import { SSHSession, SSHChannel } from './ssh_session'
+import { ControlMessage } from './webrtc_session'
 import { Failure } from './session'
 import { NativeBiometric } from 'capacitor-native-biometric'
 import { Capacitor } from "@capacitor/core"
@@ -93,6 +94,12 @@ export function loadCommands(shell: Shell): Map<string, Command> {
             usage: "i[nstall] [gatename]",
             execute: async args => installCMD(shell, args)
         },
+        login: {
+            name: "login",
+            help: "Login to an existing PeerBook account",
+            usage: "l[ogin]",
+            execute: async () => loginCMD(shell)
+        },
         map: {
             name: "map",
             help: "Back to the map",
@@ -113,7 +120,7 @@ export function loadCommands(shell: Shell): Map<string, Command> {
         },
         subscribe: {
             name: "subscribe",
-            help: "Subscripte to peerbook",
+            help: "Subscripte to PeerBook",
             usage: "sub[scribe]",
             execute: async () => subscribeCMD(shell)
         },
@@ -536,15 +543,9 @@ async function editCMD(shell:Shell, args: string[]) {
             }
             fFields = fFields.filter((_, i) => enabled[i])
             res = await shell.runForm(fFields, "text")
-            if (isPB && enabled[0]) {
-                await terminal7.pb.adminCommand({
-                    type: "rename",
-                    args: {
-                       target: gate.fp,
-                       name: res[0]
-                    }
-                })
-            }
+            if (isPB && enabled[0])
+                await terminal7.pb.adminCommand(
+                    new ControlMessage("rename", { target: gate.fp, name: res[0]}))
             gateAttrs.filter((_, i) => enabled[i])
                      .forEach((k, i) =>  {
                          let v = res[i]
@@ -564,13 +565,8 @@ async function editCMD(shell:Shell, args: string[]) {
             if (isPB) {
                 const otp = await shell.askValue("OTP")
                 try {
-                    await terminal7.pb.adminCommand({
-                        type: "delete",
-                        args: {
-                            target: gate.fp,
-                            otp: otp
-                        }
-                    })
+                    await terminal7.pb.adminCommand(
+                        new ControlMessage("delete",{ target: gate.fp, otp: otp }))
                 } catch (e) {
                     console.log("Failed to delete host", e)
                     shell.t.writeln("Failed to delete host")
@@ -838,7 +834,7 @@ export async function installCMD(shell: Shell, args: string[]) {
     let done = false
     let error = false
     const fields: Fields = [
-        { prompt: "Copy command to 📋" },
+        { prompt: "Copy command" },
         { prompt: "Cancel" },
     ]
     if (native)
@@ -850,7 +846,7 @@ export async function installCMD(shell: Shell, args: string[]) {
     const choice= await shell.runForm(fields, "menu")
     if (choice == "Cancel")
         return
-    if (choice == "Copy command to 📋") {
+    if (choice == "Copy command") {
         Clipboard.write({ string: cmd })
         shell.t.writeln("Command copied to clipboard")
         return